Drop Golint + updates + drop caching
The build is failing because `go get` is no longer supported outside of
a Go module, and we were using that to get Golint. Golint is deprecated
nowadays, so just drop it entirely.

Upgrade `setup-go` and `checkout` actions to v3.

Remove manual caching -- `setup-go` now caches Go dependencies
automatically so it's not doing much for us anymore.